add 50 products in seeders for charge database
