+++ index.js
@@ -1,8 +1,18 @@
 const app = require("./src/app");
 const { conn } = require("./src/db.js");
+const { Products } = require("./src/db");
+const seeders = require("./seeders/seeders.json");
 
-conn.sync({ force: false }).then(() => {
+const fillDataBase = () => {
+	seeders.products.forEach((product) => {
+		// esta funcion carga la base de datos
+		Products.create(product);
+	});
+};
+
+conn.sync({ force: true }).then(() => {
 	app.listen(3001, () => {
+		fillDataBase();
 		console.log("listening at 3001");
 	});
 });
