+++ src/handlers/products/getHandlerId.js
@@ -1,13 +1,15 @@
-const { getProductsId } = require("../../controllers/productsControl");
+const {
+	getProductsId,
+} = require("../../controllers/productController/productsControl");
 
 const getProductsById = async (req, res) => {
-  const { id } = req.params;
-  try {
-    const result = await getProductsId(id);
-    res.status(200).json(result);
-  } catch (error) {
-    res.status(400).send(error.message);
-  }
+	const { id } = req.params;
+	try {
+		const result = await getProductsId(id);
+		res.status(200).json(result);
+	} catch (error) {
+		res.status(400).send(error.message);
+	}
 };
 
 module.exports = getProductsById;
