--- src/handlers/products/seeders/seeders.json
@@ -1,28 +1,0 @@
-{
-	"products": [
-		{
-			"name": "Noxpirin",
-			"price": "$ 4000",
-			"description": "Medicamento para la gripa",
-			"image": "https://www.bing.com/images/search?view=detailV2&ccid=D4SKMxuM&id=E49AD8432B12132479B502AD0EECFB123E3B4B27&thid=OIP.D4SKMxuMOXhUDHejMjeWrAHaHa&mediaurl=https%3a%2f%2fortopedicosfuturo.com%2fuploads%2fcatalog%2fimages%2fd%2fe%2fp%2flarge-81000534_Noxpirin_Sinus_200mg.jpg&cdnurl=https%3a%2f%2fth.bing.com%2fth%2fid%2fR.0f848a331b8c3978540c77a3323796ac%3frik%3dJ0s7PhL77A6tAg%26pid%3dImgRaw%26r%3d0&exph=750&expw=750&q=noxpirin&simid=608027765987769160&FORM=IRPRST&ck=8560660D685E7EFF2BEAFDB82062A656&selectedIndex=0"
-		},
-		{
-			"name": "Acetaminofen",
-			"price": "$ 3000",
-			"description": "Medicamento para el dolor en general",
-			"image": "https://www.bing.com/images/search?view=detailV2&ccid=QrFb2upv&id=6CB2A3CFC07ADFF3D44ADD9BA384DBFD46F91B81&thid=OIP.QrFb2upvmcOfkldShiMbNQHaHa&mediaurl=https%3a%2f%2fwww.somoselsalvador.online%2fwp-content%2fuploads%2f2021%2f03%2facetaminofen-mk-x-100.png&cdnurl=https%3a%2f%2fth.bing.com%2fth%2fid%2fR.42b15bdaea6f99c39f92575286231b35%3frik%3dgRv5Rv3bhKOb3Q%26pid%3dImgRaw%26r%3d0&exph=1200&expw=1200&q=acetaminofen&simid=607988544352954587&FORM=IRPRST&ck=C0255831040D27721B8DD99CD542C631&selectedIndex=0"
-		},
-		{
-			"name": "Ibuprofeno",
-			"price": "$ 2000",
-			"description": "Medicamento para el dolor de cabeza y la inflamacion",
-			"image": "https://www.bing.com/images/search?view=detailV2&ccid=X3bIh0nD&id=95BC053BEDB69D9D77EF76632C1E64A04841CA88&thid=OIP.X3bIh0nDv8aK2W7Ae8B-3wHaHa&mediaurl=https%3a%2f%2fpharmahorro.com%2fwp-content%2fuploads%2f2020%2f05%2fIbuprofeno-600-mg.jpg&cdnurl=https%3a%2f%2fth.bing.com%2fth%2fid%2fR.5f76c88749c3bfc68ad96ec07bc07edf%3frik%3diMpBSKBkHixjdg%26pid%3dImgRaw%26r%3d0&exph=500&expw=500&q=ibuprofeno&simid=608042287230687502&FORM=IRPRST&ck=D933076D51CAD1D2B0E395CF0A541836&selectedIndex=5"
-		},
-		{
-			"name": "Ibuprofeno",
-			"price": "$ 2000",
-			"description": "Medicamento para el dolor de cabeza y la inflamacion",
-			"image": "https://www.bing.com/images/search?view=detailV2&ccid=X3bIh0nD&id=95BC053BEDB69D9D77EF76632C1E64A04841CA88&thid=OIP.X3bIh0nDv8aK2W7Ae8B-3wHaHa&mediaurl=https%3a%2f%2fpharmahorro.com%2fwp-content%2fuploads%2f2020%2f05%2fIbuprofeno-600-mg.jpg&cdnurl=https%3a%2f%2fth.bing.com%2fth%2fid%2fR.5f76c88749c3bfc68ad96ec07bc07edf%3frik%3diMpBSKBkHixjdg%26pid%3dImgRaw%26r%3d0&exph=500&expw=500&q=ibuprofeno&simid=608042287230687502&FORM=IRPRST&ck=D933076D51CAD1D2B0E395CF0A541836&selectedIndex=5"
-		}
-	]
-}
